Introduce AddAction
